// src/app/models/chamados/PutChamadosModel.ts
export class PutChamadosModel{
  
    id_chamados_clientes: string = '';
    nome_contato: string = '';
    email: string = '';
    telefone: string = '';  
	empresa: string = '';
	cnpj: string = ''; 
    tipo_servico: string = ''; 
    corpo_assunto: string = ''; 
    status_chamado: string = ''; 
    resposta_chamado: string = ''; 

}